reorganized event and stat DTOs into their own files and moved timeline bucketing into SQL
