+++ code/backend/Cleanuparr.Infrastructure/Stats/ArrInstanceHealthDto.cs
@@ -1,0 +1,11 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class ArrInstanceHealthDto
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Type { get; set; } = string.Empty;
+    public bool IsHealthy { get; set; }
+    public DateTimeOffset LastChecked { get; set; }
+    public string? ErrorMessage { get; set; }
+}
