+++ code/backend/Cleanuparr.Infrastructure/Stats/EventV2Stats.cs
@@ -1,0 +1,10 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class EventV2Stats
+{
+    public int TotalCount { get; set; }
+
+    public Dictionary<string, int> ByType { get; set; } = new();
+
+    public Dictionary<string, int> BySeverity { get; set; } = new();
+}
