+++ code/backend/Cleanuparr.Infrastructure/Stats/HealthStats.cs
@@ -1,0 +1,8 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class HealthStats
+{
+    public List<DownloadClientHealthDto> DownloadClients { get; set; } = [];
+
+    public List<ArrInstanceHealthDto> ArrInstances { get; set; } = [];
+}
