+++ code/backend/Cleanuparr.Infrastructure/Stats/JobStats.cs
@@ -1,0 +1,8 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class JobStats
+{
+    public Dictionary<string, JobTypeStats> ByType { get; set; } = new();
+
+    public int TimeframeHours { get; set; }
+}
