+++ code/backend/Cleanuparr.Infrastructure/Stats/MalwareV2Stats.cs
@@ -1,0 +1,6 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class MalwareV2Stats
+{
+    public int Blocked { get; set; }
+}
