+++ code/backend/Cleanuparr.Infrastructure/Stats/RecentStrikeDto.cs
@@ -1,0 +1,10 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class RecentStrikeDto
+{
+    public Guid Id { get; set; }
+    public string Type { get; set; } = string.Empty;
+    public DateTimeOffset CreatedAt { get; set; }
+    public string DownloadId { get; set; } = string.Empty;
+    public string Title { get; set; } = string.Empty;
+}
