+++ code/backend/Cleanuparr.Infrastructure/Stats/StrikeV2Stats.cs
@@ -1,0 +1,12 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class StrikeV2Stats
+{
+    public Dictionary<string, int> Active { get; set; } = new();
+
+    public int Issued { get; set; }
+
+    public int Recovered { get; set; }
+
+    public int Removed { get; set; }
+}
