+++ code/backend/Cleanuparr.Infrastructure/Stats/TimelineBucketDto.cs
@@ -1,0 +1,8 @@
+namespace Cleanuparr.Infrastructure.Stats;
+
+public class TimelineBucketDto
+{
+    public DateOnly Date { get; set; }
+
+    public int Count { get; set; }
+}
